aider: Added console.log to debug innerHTML update in Firefox.
# Aider chat conversation:

USER: this didnt make a difference

ASSISTANT: I see. If the changes didn't make a difference, it's possible that the issue lies elsewhere.

Since you mentioned that the stats are displayed in Chrome but not in Firefox, and there are no errors in the Firefox console, it might be a browser-specific issue.

One possibility is that Firefox is not rendering the updated HTML for some reason. To check if this is the case, you can add a `console.log` statement after the line where you update the `innerHTML` of the `statsElement`. This will print the current `innerHTML` of the `statsElement` to the console, allowing you to see if the HTML has been updated correctly.

Here is the *SEARCH/REPLACE* block:

client/main.js
```javascript
<<<<<<< SEARCH
        statsElement.innerHTML = statsHtml;
=======
        statsElement.innerHTML = statsHtml;
        console.log(statsElement.innerHTML);
>>>>>>> REPLACE
```

Please apply this change and check the Firefox console after you submit the form. If the `innerHTML` of the `statsElement` is printed correctly, it means that the HTML is being updated correctly, but Firefox is not rendering the updated HTML for some reason. If the `innerHTML` is not printed correctly, it means that there is an issue with the code that updates the `innerHTML`.